Variadic: org: more sequential/inviting order
Main idea is that:

Writing code I already several times wrote function `variadic`, that
converts the functions to variadic ones.

But doing it through type classes only is constraining, it is probably (almost)
not possible to make general `variadic` maker that accepts any type of function.

That is the main path I always come to variadic stuff. Questioning: is the
universal `variadic fun` possible.

So I come over to the library & look for function that convers functions to
variadic.

`ToVariadic` is the class that has `toVariadic` which are created to do that.

Also `To` and then `From` conversions is generally a good way to sequence stuff.

Put utlitarian families after the class, well, again - because people don't need
to see or be greeted with utilitarian preparation, to spair the details.